+++ unity/Runtime/Bindings/MjBindings.cs
@@ -678,10 +678,11 @@ public enum mjtRndFlag : int{
   mjRND_SKYBOX = 4,
   mjRND_FOG = 5,
   mjRND_HAZE = 6,
-  mjRND_SEGMENT = 7,
-  mjRND_IDCOLOR = 8,
-  mjRND_CULL_FACE = 9,
-  mjNRNDFLAG = 10,
+  mjRND_DEPTH = 7,
+  mjRND_SEGMENT = 8,
+  mjRND_IDCOLOR = 9,
+  mjRND_CULL_FACE = 10,
+  mjNRNDFLAG = 11,
 }
 public enum mjtStereo : int{
   mjSTEREO_NONE = 0,
@@ -6315,7 +6316,7 @@ public unsafe struct mjvScene_ {
   public fixed float rotate[4];
   public float scale;
   public int stereo;
-  public fixed byte flags[10];
+  public fixed byte flags[11];
   public int framewidth;
   public fixed float framergb[3];
   public int status;