+++ doc/changelog.rst
@@ -12,6 +12,7 @@ General
   - The type of the ``sig`` (signature) argument of :ref:`mj_stateSize` and related functions has been changed from
     ``unsigned int`` to ``int``. Before this change, invalid negative arguments passed to this function would result in
     a silent implicit cast, now negativity will trigger an error.
+  - Added a :ref:`depth<mjtRndFlag>` rendering flag
 
 MJX
 ^^^